requirements.txt
astroid==2.15.0
boto3==1.26.94
botocore==1.29.94
dnspython==2.3.0
docutils==0.19
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
pylint==2.17.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
urllib3==1.26.15
wrapt==1.15.0
